+++ src/mock/filter.js
@@ -1,12 +1,18 @@
+import {getRandomArrayItem} from "../utils";
+
 const filterNames = [
   `all`, `overdue`, `today`, `favorites`, `repeating`, `archive`
+];
+
+const filterValues = [
+  `11`, `69`, `13`, `7`, `42`, `15`, `34`
 ];
 
 const generateFilters = () => {
   return filterNames.map((it) => {
     return {
       name: it,
-      count: Math.floor(Math.random() * 15),
+      count: getRandomArrayItem(filterValues)
     };
   });
 };
