+++ src/mock/task.js
@@ -1,4 +1,6 @@
 import {COLORS} from "../consts";
+import {getRandomArrayItem, getRandomDate} from "../utils";
+
 
 const DefaultRepeatingDays = {
   "mo": false,
@@ -16,25 +18,7 @@ const DescriptionItems = [
   `Пройти интенсив на соточку`,
 ];
 
-const getRandomIntegerNumber = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min));
-};
 
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length);
-
-  return array[randomIndex];
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomIntegerNumber(0, 8);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
-};
 
 const generateRepeatingDays = () => {
   return Object.assign({}, DefaultRepeatingDays, {
